requirements.txt
flask>=3.0.3
gunicorn==20.0.4
flask-cors==5.0.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
